Load raw images and transform(no for cammera mode
git-tfs-id: [https://devops.cyberoptics.com/TPC1]$/CyberStitch;C48641

+++ Main/src/logic/MosaicDataModel/MosaicSet.cpp
@@ -241,7 +241,7 @@ namespace MosaicDM
 				deviceIndice.push_back(iIndex);
 		}
 
-		return(deviceIndice.size());
+		return((int)deviceIndice.size());
 	}
 
 	bool MosaicSet::HasAllImages()
